Separate history management into its own object
- the object also manages unflushed history
- the history DB has its own version, starting at 0

This is the first step to making history management into
a separate service and to larger block sizes.

The next step is an improved history format, which I hope will both
save space and be a solution to issue #185.  The DB should be able to
upgrade in-place without re-syncing the chain.